+++ internal/tests/SchemaVersionResource/schema_version_resource_avro_test.go
@@ -63,7 +63,7 @@ func TestSchemaVersionAvroResource(t *testing.T) {
 			{
 				// To ensure cleanup if one of the test cases had an error
 				Destroy: true,
-				Config:  GetProvider() + GetFile("axual_schema_version_avro_v3_replaced.tf"),
+				Config:  GetProvider() + GetFile("axual_schema_version_avro_multiple_versions_for_same_schema.tf"),
 			},
 		},
 	})
@@ -91,7 +91,7 @@ func TestSchemaVersionAvroWithOwnersResource(t *testing.T) {
 			{
 				// To ensure cleanup if one of the test cases had an error
 				Destroy: true,
-				Config:  GetProvider() + GetFile("axual_schema_version_avro_v3_replaced.tf"),
+				Config:  GetProvider() + GetFile("axual_schema_version_avro_multiple_versions_for_same_schema.tf"),
 			},
 		},
 	})