+++ axual-webclient/environment_data.go
@@ -211,5 +211,5 @@ type EnvironmentRequest struct {
 	Owners              string                 `json:"owners,omitempty"`
 	Viewers             []string               `json:"viewers"`
 	Properties          map[string]interface{} `json:"properties,omitempty"`
-	Settings            map[string]interface{} `json:"settings,omitempty"`
+	Settings            map[string]interface{} `json:"settings"`
 }