+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [master](https://github.com/Axual/terraform-provider-axual/blob/master) - TBR
+* Added schema owner to the SchemaVersion resource and datasource
+* Added settings field to environment resource
+
 ## [2.4.2](https://github.com/Axual/terraform-provider-axual/releases/tag/v2.4.2) - 2024-12-12
 * Added `Instance` data source.
 * Added error handling for environment, group and topic data sources.